The Cover Songs Dataset! (with Graphs attached)
I've decided to add the Cover Songs project as a markdown file (.md) since Github is having trouble rendering Jupyter Notebooks.
The project consists of Webscraping information from a URL that contains a website where users can submit Cover songs of popular artists. I've decided to analyze  and visualize this information. Therefore, I've attached 2 images of the Graphs which can't be seen on a markdown file.